Fixed MSVC compile warnings. Discovered issue with delayed adding of space.
- Need to delay initialization of components until next update, as game_hack component is created before engine is run and thus
     before spaces are added. Because of this it cannot access the perf_vis space.

+++ src/LE/engine/tilemap_component.cpp
@@ -31,7 +31,7 @@ tileset::tileset(std::string const& tsd_file_name)
   p_num_tiles = num_tiles.x() * num_tiles.y();
 
   // Dimensions of a single tile when entire tileset is considered using coordinates [0, 1]
-  vec2 const tile_tex_coord_dim = vec2(p_tile_size) / vec2(texture_dim);
+  vec2 const tile_tex_coord_dim = vec2((float)p_tile_size) / vec2(texture_dim);
 
   uvec2 curr_tile;
   for(curr_tile.y() = 0; curr_tile.y() < num_tiles.y(); ++curr_tile.y())
